+++ src/Symfony/Component/Security/Http/Tests/EventListener/PasswordMigratingListenerTest.php
@@ -65,11 +65,21 @@ class PasswordMigratingListenerTest extends TestCase
 
         // blank password
         yield [$this->createEvent(new SelfValidatingPassport(new UserBadge('test', function () { return $this->createMock(TestPasswordAuthenticatedUser::class); }), [new PasswordUpgradeBadge('', $this->createPasswordUpgrader())]))];
-
-        // no user
-        yield [$this->createEvent($this->createMock(PassportInterface::class))];
     }
 
+    /**
+     * @group legacy
+     */
+    public function testLegacyUnsupportedEvents()
+    {
+        $this->hasherFactory->expects($this->never())->method('getPasswordHasher');
+
+        $this->listener->onLoginSuccess($this->createEvent($this->createMock(PassportInterface::class)));
+    }
+
+    /**
+     * @group legacy
+     */
     public function testUnsupportedPassport()
     {
         // A custom Passport, without an UserBadge