+++ src/Symfony/Component/Form/CHANGELOG.md
@@ -21,6 +21,7 @@ CHANGELOG
 5.4
 ---
 
+ * Deprecate calling `FormErrorIterator::children()` if the current element is not iterable.
  * Allow to pass `TranslatableMessage` objects to the `help` option
 
 5.3