+++ src/Symfony/Component/Security/Http/CHANGELOG.md
@@ -16,7 +16,7 @@ CHANGELOG
  * Deprecate `DeauthenticatedEvent`, use `TokenDeauthenticatedEvent` instead
  * Deprecate `CookieClearingLogoutHandler`, `SessionLogoutHandler` and `CsrfTokenClearingLogoutHandler`.
    Use `CookieClearingLogoutListener`, `SessionLogoutListener` and `CsrfTokenClearingLogoutListener` instead
- * Deprecate `PassportInterface` and `UserPassportInterface`, use `Passport` instead
+ * Deprecate `PassportInterface`, `UserPassportInterface` and `PassportTrait`, use `Passport` instead
 
 5.3
 ---